refactor(Wish) : 컨트롤러 수정
### 변경점
- 기능적 변경 때문에 컨트롤러를 admin/user 2개로 분리
- admin쪽의 리턴값을 String으로 통일
